refactor(node-ui): break the health/coverage cycle via a pca-primitives leaf
#1349 briefly made health.ts import `hasPcaBudget` from coverage.ts while
coverage.ts imports `isPcaExpired` back — a reference cycle (safe, but a smell,
flagged in the gate). Extract the pure snapshot predicates (`isPcaExpired`,
`isPcaDead`, `pcaBudgetState`, `hasPcaBudget`, `bigGt0`) into a new leaf
`pca/pca-primitives.ts` that imports ONLY types, and point coverage.ts,
health.ts and walletBinding.ts at it. The pca DAG is acyclic again; coverage.ts
keeps the spendability composite + probe classification and its header no longer
claims to be the "PURE leaf".

Also folds in the #1349 solvency/precedence tests the gate asked for
(remainingAllowance-positive stays healthy; remainingAllowance=0 supersedes a
positive baseEpochAllowance proxy -> insolvent) and splits the leaf imports in
the coverage/components tests.
